fixed Bugs on ORIGINAL openapi-diff-ignore
and added test by version.
