+++ pom.xml
@@ -28,7 +28,7 @@
 
     <groupId>com.github.elibracha</groupId>
     <artifactId>openapi-diff-ignore</artifactId>
-    <version>2.2.8-SNAPSHOT</version>
+    <version>2.2.10-SNAPSHOT</version>
 
     <properties>
         <maven.compile.plugin.source>1.8</maven.compile.plugin.source>
