+++ samples/sample-api-first-approach/pom.xml
@@ -88,7 +88,7 @@
         <dependency>
             <groupId>com.github.elibracha</groupId>
             <artifactId>openapi-diff</artifactId>
-            <version>2.2.10</version>
+            <version>2.2.10-SNAPSHOT</version>
             <scope>test</scope>
             <exclusions>
                 <exclusion>
