+++ samples/sample-api-first-approach/src/test/resources/openapi/diff/api_diff-v1.html
@@ -1,1 +1,1 @@
-<!DOCTYPE html><html lang="en"><head><meta charset="utf-8"><title>API Change Log</title><link rel="stylesheet" href="file:/home/elibracha/.m2/repository/com/github/elibracha/openapi-diff/2.2.10/openapi-diff-2.2.10.jar!/demo.css"></head><body><header><h1>API Change Log</h1></header><div class="article"><div><h2>What&#x27;s New</h2><hr><ol></ol></div><div><h2>What&#x27;s Deleted</h2><hr><ol></ol></div><div><h2>What&#x27;s Deprecated</h2><hr><ol></ol></div><div><h2>What&#x27;s Changed</h2><hr><ol></ol></div></div></body></html>
+<!DOCTYPE html><html lang="en"><head><meta charset="utf-8"><title>Api Change Log</title><link rel="stylesheet" href="http://deepoove.com/swagger-diff/stylesheets/demo.css"></head><body><header><h1>Api Change Log</h1></header><div class="article"><div><h2>What&#x27;s New</h2><hr><ol></ol></div><div><h2>What&#x27;s Deleted</h2><hr><ol></ol></div><div><h2>What&#x27;s Deprecated</h2><hr><ol></ol></div><div><h2>What&#x27;s Changed</h2><hr><ol></ol></div></div></body></html>
